test world mesh update

+++ gtools/gui/lib/tile_renderer.py
@@ -15,7 +15,7 @@ from gtools.gui.camera import Camera2D
 from gtools.gui.camera3d import Camera3D
 from gtools.gui.lib import layer
 from gtools.gui.lib.renderer import Renderer
-from gtools.gui.lib.tree_renderer import TreeRenderer
+from gtools.gui.lib.tree_renderer import TreeMesh, TreeRenderer
 from gtools.gui.opengl import Mesh, ShaderProgram, Uniform
 from gtools.gui.texture import GLTexManager, TextureArray
 
@@ -70,8 +70,12 @@ class TileRenderer(Renderer):
         self._opacity3d = self._shader3d.get_uniform("u_opacity")
 
         self._tree_renderer = TreeRenderer()
+        self.tree_mesh: TreeMesh | None = None
+        self._chunk_meshes: dict[tuple[int, int], list[tuple[str, TextureArray, Mesh]]] = {}
+        self.CHUNK_SIZE = 8
 
     def load(self, world: World) -> None:
+        self.delete()
         self._build_meshes(world)
         self._tex_mgr.flush()
 
@@ -85,7 +89,8 @@ class TileRenderer(Renderer):
         self._shader.use()
         self._mvp.set_mat4x4(camera.proj_as_numpy())
         self._draw_layers(camera, self._tex, self._layer, self._opacity)
-        self._tree_renderer.draw(camera, self.tree_mesh)
+        if self.tree_mesh:
+            self._tree_renderer.draw(camera, self.tree_mesh)
 
     def draw_3d(self, camera3d: Camera3D, layer_spread: float) -> None:
         if not self.any():
@@ -95,15 +100,20 @@ class TileRenderer(Renderer):
         self._vp3d.set_mat4x4(camera3d.view_proj_as_numpy())
         self._spread3d.set_float(layer_spread)
         self._draw_layers(None, self._tex3d, self._layer3d, self._opacity3d)
-        self._tree_renderer.draw_3d(camera3d, self.tree_mesh, layer_spread)
+        if self.tree_mesh:
+            self._tree_renderer.draw_3d(camera3d, self.tree_mesh, layer_spread)
 
     def delete(self) -> None:
-        self.tree_mesh.delete()
-        self._tree_renderer.delete()
+        if self.tree_mesh:
+            self.tree_mesh.delete()
+            self.tree_mesh = None
+
+        for chunk_list in self._chunk_meshes.values():
+            for _, _, mesh in chunk_list:
+                mesh.delete()
+        self._chunk_meshes.clear()
+
         for rl in self._layers.values():
-            for chunk_list in rl.chunks.values():
-                for _, mesh in chunk_list:
-                    mesh.delete()
             rl.chunks.clear()
 
     def _draw_layers(self, camera: Camera2D | None, tex_uniform: Uniform, layer_uniform: Uniform, opacity: Uniform) -> None:
@@ -127,133 +137,186 @@ class TileRenderer(Renderer):
             if not rl.depth_write:
                 glDepthMask(GL_TRUE)
 
-    def _build_meshes(self, world: World) -> None:
-        CHUNK_SIZE = 20
-        # layer -> (chunk_x, chunk_y) -> tex_array -> list[float]
-        instances: dict[str, dict[tuple[int, int], dict[TextureArray, list[float]]]] = {
-            key: defaultdict(lambda: defaultdict(list)) for key in self._layers
-        }
-        trees: list[Tile] = []
+    def update_tile(self, x: int, y: int, world: World) -> None:
+        chunk_x = x // self.CHUNK_SIZE
+        chunk_y = y // self.CHUNK_SIZE
+        chunk_key = (chunk_x, chunk_y)
 
+        self.delete_chunk(chunk_key)
+        self._build_chunk(world, chunk_x, chunk_y)
+
+        # check for trees in the area
+        # TODO: optimize this to only rebuild trees in the chunk
+        # but trees are complex because they span multiple tiles (visually) 
+        # and TreeRenderer currently rebuilds everything.
+        trees = []
         for tile in world.tiles.values():
-            chunk_x = int(tile.pos.x // CHUNK_SIZE)
-            chunk_y = int(tile.pos.y // CHUNK_SIZE)
-            chunk_key = (chunk_x, chunk_y)
+            if tile.fg_id and tile.extra and isinstance(tile.extra, SeedTile):
+                trees.append(tile)
 
-            item = item_database.get(tile.fg_id)
+        if self.tree_mesh:
+            self.tree_mesh.delete()
+        self.tree_mesh = self._tree_renderer.build(trees)
+        self._tex_mgr.flush()
 
-            if tile.bg_id:
-                tex_array, data = self._tile_instance_data(tile, tile.bg_id, tile.bg_tex_index)
-                instances["bg"][chunk_key][tex_array].extend(data)
+    def delete_chunk(self, chunk_key: tuple[int, int]) -> None:
+        if chunk_key not in self._chunk_meshes:
+            return
 
-            if tile.fg_id:
-                handled = False
+        for layer_name, tex_array, mesh in self._chunk_meshes[chunk_key]:
+            rl = self._layers[layer_name]
+            if tex_array in rl.chunks:
+                rl.chunks[tex_array] = [x for x in rl.chunks[tex_array] if x[1] != mesh]
+            mesh.delete()
 
-                if tile.extra:
-                    if isinstance(tile.extra, DisplayBlockTile):
+        del self._chunk_meshes[chunk_key]
+
+    def _build_meshes(self, world: World) -> None:
+        self.delete()
+
+        chunk_coords = set()
+        for tile in world.tiles.values():
+            chunk_coords.add((tile.pos.x // self.CHUNK_SIZE, tile.pos.y // self.CHUNK_SIZE))
+
+        for cx, cy in chunk_coords:
+            self._build_chunk(world, cx, cy)
+
+        trees = [t for t in world.tiles.values() if t.fg_id and t.extra and isinstance(t.extra, SeedTile)]
+        self.tree_mesh = self._tree_renderer.build(trees)
+
+    def _build_chunk(self, world: World, chunk_x: int, chunk_y: int) -> None:
+        # layer -> tex_array -> list[float]
+        instances: dict[str, dict[TextureArray, list[float]]] = {
+            key: defaultdict(list) for key in self._layers
+        }
+
+        start_x = chunk_x * self.CHUNK_SIZE
+        start_y = chunk_y * self.CHUNK_SIZE
+        end_x = start_x + self.CHUNK_SIZE
+        end_y = start_y + self.CHUNK_SIZE
+
+        for y in range(start_y, end_y):
+            for x in range(start_x, end_x):
+                tile = world.get_tile(x, y)
+                if not tile:
+                    continue
+
+                item = item_database.get(tile.fg_id)
+
+                if tile.bg_id:
+                    tex_array, data = self._tile_instance_data(tile, tile.bg_id, tile.bg_tex_index)
+                    instances["bg"][tex_array].extend(data)
+
+                if tile.fg_id:
+                    handled = False
+
+                    if tile.extra:
+                        if isinstance(tile.extra, DisplayBlockTile):
+                            tex_array, data = self._tile_instance_data_raw(
+                                tile,
+                                item.texture_file.decode(),
+                                ivec2(item.tex_coord_x, item.tex_coord_y + 1),
+                            )
+                            instances["fg_before"][tex_array].extend(data)
+
+                        elif isinstance(tile.extra, VendingMachineTile):
+                            tex = 0
+                            if tile.extra.price == 0 and tile.extra.item_id == 0:
+                                tex = 1
+
+                            tex_array, data = self._tile_instance_data_raw(
+                                tile,
+                                item.texture_file.decode(),
+                                ivec2(item.tex_coord_x + tex, item.tex_coord_y),
+                            )
+                            instances["fg"][tex_array].extend(data)
+
+                            if tile.flags & TileFlags.FG_ALT_MODE:  # has wl inside
+                                tex_array, data = self._tile_instance_data_raw(
+                                    tile,
+                                    item.texture_file.decode(),
+                                    ivec2(item.tex_coord_x + 3, item.tex_coord_y),
+                                )
+                                instances["fg_after"][tex_array].extend(data)
+
+                            if tile.extra.price == 0 or tile.extra.item_id == 0:
+                                tex_array, data = self._tile_instance_data_raw(
+                                    tile,
+                                    item.texture_file.decode(),
+                                    ivec2(item.tex_coord_x + 2, item.tex_coord_y),
+                                )
+                                instances["fg_after"][tex_array].extend(data)
+
+                            handled = True
+                        elif isinstance(tile.extra, SeedTile):
+                            handled = True
+
+                    if item.is_steam():
+                        anchor = item_database.get(STEAM_TUBES)
+                        tex_pos = ivec2(anchor.tex_coord_x + 1, anchor.tex_coord_y)
+                        stride = get_tex_stride(ItemInfoTextureType.SMART_EDGE)
+                        off = ivec2(tile.fg_tex_index % max(stride, 1), tile.fg_tex_index // stride if stride else 0)
+
+                        tex_array, data = self._tile_instance_data_raw(
+                            tile,
+                            anchor.texture_file.decode(),
+                            tex_pos + off
+                        )
+                        instances["fg_after"][tex_array].extend(data)
+                    elif tile.fg_id == STEAM_REVOLVER:
                         tex_array, data = self._tile_instance_data_raw(
                             tile,
                             item.texture_file.decode(),
                             ivec2(item.tex_coord_x, item.tex_coord_y + 1),
                         )
-                        instances["fg_before"][chunk_key][tex_array].extend(data)
+                        instances["fg_after"][tex_array].extend(data)
 
-                    elif isinstance(tile.extra, VendingMachineTile):
-                        tex = 0
-                        if tile.extra.price == 0 and tile.extra.item_id == 0:
-                            tex = 1
+                    if not handled:
+                        tex_array, data = self._tile_instance_data(tile, tile.fg_id, tile.fg_tex_index)
+                        instances["fg"][tex_array].extend(data)
 
-                        tex_array, data = self._tile_instance_data_raw(
-                            tile,
-                            item.texture_file.decode(),
-                            ivec2(item.tex_coord_x + tex, item.tex_coord_y),
-                        )
-                        instances["fg"][chunk_key][tex_array].extend(data)
-
-                        if tile.flags & TileFlags.FG_ALT_MODE:  # has wl inside
-                            tex_array, data = self._tile_instance_data_raw(
-                                tile,
-                                item.texture_file.decode(),
-                                ivec2(item.tex_coord_x + 3, item.tex_coord_y),
-                            )
-                            instances["fg_after"][chunk_key][tex_array].extend(data)
-
-                        if tile.extra.price == 0 or tile.extra.item_id == 0:
-                            tex_array, data = self._tile_instance_data_raw(
-                                tile,
-                                item.texture_file.decode(),
-                                ivec2(item.tex_coord_x + 2, item.tex_coord_y),
-                            )
-                            instances["fg_after"][chunk_key][tex_array].extend(data)
-
-                        handled = True
-                    elif isinstance(tile.extra, SeedTile):
-                        trees.append(tile)
-                        handled = True
-
-                if item.is_steam():
-                    anchor = item_database.get(STEAM_TUBES)
-                    tex_pos = ivec2(anchor.tex_coord_x + 1, anchor.tex_coord_y)
+                if tile.flags & TileFlags.ON_FIRE:
                     stride = get_tex_stride(ItemInfoTextureType.SMART_EDGE)
-                    off = ivec2(tile.fg_tex_index % max(stride, 1), tile.fg_tex_index // stride if stride else 0)
+                    tex_index = tile.overlay_tex_index
+                    tex_pos = ivec2(tex_index % max(stride, 1), tex_index // stride if stride else 0)
 
-                    tex_array, data = self._tile_instance_data_raw(
-                        tile,
-                        anchor.texture_file.decode(),
-                        tex_pos + off
-                    )
-                    instances["fg_after"][chunk_key][tex_array].extend(data)
-                elif tile.fg_id == STEAM_REVOLVER:
-                    tex_array, data = self._tile_instance_data_raw(
-                        tile,
-                        item.texture_file.decode(),
-                        ivec2(item.tex_coord_x, item.tex_coord_y + 1),
-                    )
-                    instances["fg_after"][chunk_key][tex_array].extend(data)
+                    tex_array, data = self._tile_instance_data_raw(tile, "fire.rttex", tex_pos)
+                    instances["fire"][tex_array].extend(data)
+                elif tile.flags & TileFlags.IS_WET:
+                    stride = get_tex_stride(ItemInfoTextureType.SMART_EDGE)
+                    tex_index = tile.overlay_tex_index
+                    tex_pos = ivec2(tex_index % max(stride, 1), tex_index // stride if stride else 0)
 
-                if not handled:
-                    tex_array, data = self._tile_instance_data(tile, tile.fg_id, tile.fg_tex_index)
-                    instances["fg"][chunk_key][tex_array].extend(data)
+                    tex_array, data = self._tile_instance_data_raw(tile, "water.rttex", tex_pos)
+                    instances["water"][tex_array].extend(data)
 
-            if tile.flags & TileFlags.ON_FIRE:
-                stride = get_tex_stride(ItemInfoTextureType.SMART_EDGE)
-                tex_index = tile.overlay_tex_index
-                tex_pos = ivec2(tex_index % max(stride, 1), tex_index // stride if stride else 0)
+        bounds = (
+            chunk_x * self.CHUNK_SIZE * self.TILE_SIZE - 16,
+            chunk_y * self.CHUNK_SIZE * self.TILE_SIZE - 16,
+            self.CHUNK_SIZE * self.TILE_SIZE,
+            self.CHUNK_SIZE * self.TILE_SIZE,
+        )
 
-                tex_array, data = self._tile_instance_data_raw(tile, "fire.rttex", tex_pos)
-                instances["fire"][chunk_key][tex_array].extend(data)
-            elif tile.flags & TileFlags.IS_WET:
-                stride = get_tex_stride(ItemInfoTextureType.SMART_EDGE)
-                tex_index = tile.overlay_tex_index
-                tex_pos = ivec2(tex_index % max(stride, 1), tex_index // stride if stride else 0)
+        chunk_key = (chunk_x, chunk_y)
+        self._chunk_meshes[chunk_key] = []
 
-                tex_array, data = self._tile_instance_data_raw(tile, "water.rttex", tex_pos)
-                instances["water"][chunk_key][tex_array].extend(data)
-
-        self.tree_mesh = self._tree_renderer.build(trees)
-
-        for key, chunks_data in instances.items():
-            rl = self._layers[key]
-            for (cx, cy), per_tex in chunks_data.items():
-                bounds = (
-                    cx * CHUNK_SIZE * self.TILE_SIZE - 16,
-                    cy * CHUNK_SIZE * self.TILE_SIZE - 16,
-                    CHUNK_SIZE * self.TILE_SIZE,
-                    CHUNK_SIZE * self.TILE_SIZE,
+        for layer_key, per_tex in instances.items():
+            rl = self._layers[layer_key]
+            for tex_array, inst_list in per_tex.items():
+                if not inst_list:
+                    continue
+                mesh = Mesh(
+                    Mesh.RECT_WITH_UV_VERTS,
+                    TileRenderer.LAYOUT,
+                    Mesh.RECT_INDICES,
+                    instance_data=np.array(inst_list, dtype=np.float32),
+                    instance_layout=TileRenderer.INSTANCE_LAYOUT,
+                    instance_attrib_base=3,
                 )
-
-                for tex_array, inst_list in per_tex.items():
-                    mesh = Mesh(
-                        Mesh.RECT_WITH_UV_VERTS,
-                        TileRenderer.LAYOUT,
-                        Mesh.RECT_INDICES,
-                        instance_data=np.array(inst_list, dtype=np.float32),
-                        instance_layout=TileRenderer.INSTANCE_LAYOUT,
-                        instance_attrib_base=3,
-                    )
-                    if tex_array not in rl.chunks:
-                        rl.chunks[tex_array] = []
-                    rl.chunks[tex_array].append((bounds, mesh))
+                if tex_array not in rl.chunks:
+                    rl.chunks[tex_array] = []
+                rl.chunks[tex_array].append((bounds, mesh))
+                self._chunk_meshes[chunk_key].append((layer_key, tex_array, mesh))
 
     def _tile_instance_data(self, tile: Tile, item_id: int, tex_index: int) -> tuple[TextureArray, list[float]]:
         item = item_database.get(item_id)
